src/test/java/com/leetcode/strings/easy/ValidPalindrome_680Test.java
package com.leetcode.strings.easy;

import org.junit.jupiter.api.extension.ExtensionContext;
import org.junit.jupiter.params.ParameterizedTest;
import org.junit.jupiter.params.provider.Arguments;
import org.junit.jupiter.params.provider.ArgumentsProvider;
import org.junit.jupiter.params.provider.ArgumentsSource;

import java.util.stream.Stream;

import static org.assertj.core.api.Assertions.assertThat;
import static org.junit.jupiter.params.provider.Arguments.arguments;

class ValidPalindrome_680Test {

  private final ValidPalindrome_680 solution = new ValidPalindrome_680();

  private static class TestArgumentsProvider implements ArgumentsProvider {
    @Override
    public Stream<? extends Arguments> provideArguments(ExtensionContext context) {
      return Stream.of(arguments("aba", true), arguments("abca", true), arguments("abc", false));
    }
  }

  @ParameterizedTest
  @ArgumentsSource(TestArgumentsProvider.class)
  void validPalindrome(String str, boolean expected) {
    boolean result = solution.validPalindrome(str);

    assertThat(result).isEqualTo(expected);
  }
}
